+++ packages/veui/src/components/Select/OptionGroup.vue
@@ -2,18 +2,28 @@
 <script>
 import Option from './Option'
 import Overlay from '../Overlay'
+import Icon from '../Icon'
 import ui from '../../mixins/ui'
+import overlay from '../../mixins/overlay'
 import menu from '../../mixins/menu'
 import select from '../../mixins/select'
+import keySelect from '../../mixins/key-select'
+import outside from '../../directives/outside'
 import '../../config/uiTypes'
+import { walk } from '../../utils/data'
+import { pull } from 'lodash'
 
 export default {
   name: 'veui-option-group',
   uiTypes: ['menu'],
-  mixins: [ui, menu, select],
+  mixins: [ui, menu, select, overlay, keySelect],
   components: {
     'veui-option': Option,
-    'veui-overlay': Overlay
+    'veui-overlay': Overlay,
+    'veui-icon': Icon
+  },
+  directives: {
+    outside
   },
   props: {
     label: String,
@@ -29,7 +39,17 @@ export default {
   data () {
     return {
       items: [],
-      expanded: false
+      expanded: false,
+      localOverlayOptions: {
+        position: 'right-start',
+        constraints: [
+          {
+            to: 'window',
+            attachment: 'together'
+          }
+        ]
+      },
+      outsideRefs: ['button']
     }
   },
   computed: {
@@ -40,7 +60,7 @@ export default {
       return this.items.map(({ id }) => id)
     },
     canPopOut () {
-      return !!(this.position === 'popout' && this.options && this.options.length && this.label)
+      return !!(this.position === 'popout' && this.items && this.items.length && this.label)
     }
   },
   render () {
@@ -76,14 +96,12 @@ export default {
             }>
           </veui-option-group>
           : <veui-option
-            v-show={!option.hidden}
             {...{
               props: {
                 ...option,
                 ui: this.inheritedUi
               }
             }}
-            tabindex={option.hidden ? -1 : false}
             key={i}
             slots={
               {
@@ -95,39 +113,103 @@ export default {
       })
       : this.$slots.default
 
+    let LabelTag = this.canPopOut ? 'button' : 'div'
+
     return <div
       class={{
         'veui-option-group': true,
-        'veui-option-group-unlabelled': !this.label
+        'veui-option-group-unlabelled': !this.label,
+        'veui-option-group-expanded': this.expanded
       }}
       ui={this.ui}
       ref="label">
       {
         this.label
-          ? <div class="veui-option-group-label">
+          ? <LabelTag
+            ref="button"
+            class={{
+              'veui-option-group-label': true,
+              'veui-option-group-button': this.canPopOut
+            }}
+            {...this.canPopOut
+              ? {
+                on: {
+                  click: () => {
+                    this.expanded = true
+                  },
+                  keydown: e => {
+                    if (e.key === 'Right' || e.key === 'ArrowRight') {
+                      this.expanded = true
+                      e.stopPropagation()
+                      e.preventDefault()
+                    }
+                  }
+                }
+              }
+              : {}
+            }>
+            <span class="veui-option-label">
+              {
+                this.$scopedSlots.label
+                  ? this.$scopedSlots.label({ label: this.label })
+                  : this.label
+              }
+            </span>
             {
-              this.$scopedSlots.label
-                ? this.$scopedSlots.label({ label: this.label })
-                : this.label
+              this.canPopOut
+                ? <veui-icon class="veui-option-group-expandable" name={this.icons.expandable}/>
+                : null
             }
-          </div>
+          </LabelTag>
           : null
       }
       {
-        this.canPopOut && this.expanded
+        this.canPopOut
           ? <veui-overlay
-            target="label"
+            ref="overlay"
+            target="button"
             open={this.expanded}
-            options={{
-              position: 'right top'
-            }}
+            options={this.realOverlayOptions}
+            overlayClass={this.mergeOverlayClass('veui-option-group-box')}
             autofocus
             modal>
-            {content}
+            <div
+              ref="box"
+              class="veui-select-options"
+              tabindex="-1"
+              ui={this.ui}
+              {...{
+                directives: [{
+                  name: 'outside',
+                  value: {
+                    refs: this.outsideRefs,
+                    handler: () => {
+                      this.expanded = false
+                    }
+                  }
+                }]
+              }}
+              onKeydown={this.handleKeydown}>
+              {content}
+            </div>
           </veui-overlay>
           : content
       }
     </div>
+  },
+  watch: {
+    expanded (val) {
+      let box = this.$refs.box
+      let parent = this.menu || this.select
+      while (parent) {
+        if (val) {
+          parent.outsideRefs.push(box)
+        } else {
+          pull(parent.outsideRefs, box)
+        }
+        parent = parent.menu || parent.select
+      }
+    }
   },
   methods: {
     add (item) {
@@ -144,6 +226,47 @@ export default {
     },
     find (val) {
       return findItemByValue(this.items, val)
+    },
+    relocate () {
+      if (this.canPopOut && this.expanded) {
+        this.$refs.overlay.relocate()
+      }
+    },
+    relocateDeep () {
+      walk(this, child => {
+        if (child.$options.name === this.$options.name) {
+          child.relocate()
+        }
+      }, '$children')
+    },
+    close () {
+      this.expanded = false
+    },
+    handleKeydown (e) {
+      let passive = false
+      switch (e.key) {
+        case 'Esc':
+        case 'Escape':
+        case 'Left':
+        case 'ArrowLeft':
+          this.expanded = false
+          break
+        case 'Up':
+        case 'ArrowUp':
+          this.navigate(false)
+          break
+        case 'Down':
+        case 'ArrowDown':
+          this.navigate()
+          break
+        default:
+          passive = true
+          break
+      }
+      if (!passive) {
+        e.stopPropagation()
+        e.preventDefault()
+      }
     }
   }
 }
@@ -169,4 +292,5 @@ function findItemByValue (items, val) {
   })
   return result
 }
+
 </script>