+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 
 ### 💡 主要变更
 
+* [+] 为 `Select` 增加了 `filter` prop，用来过滤下拉内容。
+* [+] 为 `OptionGroup` 增加了 `position` prop，用来指定在弹出菜单中显示。
+* [+] 为 `Option` 增加了 `hidden` prop。
+* [+] 为 `Overlay` 增加了 `locate` 事件，在位置发生变化时触发（时机为 `tether` 的 `reposition` 事件）。
 * [^] 为 `uiTypes` 定制了选项合并策略，并修正了 `Select` 组件在 `uiTypes` 中声明的 `input` 被 mixin 中加入的 `select` 覆盖的问题。
 * [+] `Searchbox` 组件增加 `suggest-trigger` prop，用来指定推荐列表的弹出时机；增加 `suggest` 事件，当需要显示推荐列表时触发。
 * [+] `Field` 的 `rules` 中增加 `priority` 的配置，用来覆盖当前内置的规则优先级。